+++ config/site.rb
@@ -1,4 +1,4 @@
-#require 'compass'
+require 'compass'
 
 # Default is 3000
 # configuration.preview_server_port = 3000